[fix] 프로젝트(모집글) 리포, 서비스, 컨트롤러 수정 (#6)

+++ Gathering_be/src/main/java/com/Gathering_be/service/ProjectService.java
@@ -1,13 +1,15 @@
 package com.Gathering_be.service;
 
-import com.Gathering_be.domain.Member;
+import com.Gathering_be.domain.Profile;
 import com.Gathering_be.domain.Project;
 import com.Gathering_be.dto.request.ProjectCreateRequest;
 import com.Gathering_be.dto.request.ProjectUpdateRequest;
-import com.Gathering_be.dto.response.ProjectResponse;
-import com.Gathering_be.exception.MemberNotFoundException;
+import com.Gathering_be.dto.response.ProjectDetailResponse;
+import com.Gathering_be.dto.response.ProjectSimpleResponse;
+import com.Gathering_be.exception.ProfileNotFoundException;
 import com.Gathering_be.exception.ProjectNotFoundException;
-import com.Gathering_be.repository.MemberRepository;
+import com.Gathering_be.exception.UnauthorizedAccessException;
+import com.Gathering_be.repository.ProfileRepository;
 import com.Gathering_be.repository.ProjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -19,76 +21,88 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ProjectService {
-    private final MemberRepository memberRepository;
+    private final ProfileRepository profileRepository;
     private final ProjectRepository projectRepository;
 
     // 프로젝트 생성
     @Transactional
-    public Project createProject(ProjectCreateRequest request) {
+    public ProjectDetailResponse createProject(ProjectCreateRequest request) {
         Long memberId = getCurrentUserId();
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(MemberNotFoundException::new);
+        Profile profile = findProfileByMemberId(memberId);
 
         Project project = Project.builder()
-                .member(member)
+                .profile(profile)
                 .title(request.getTitle())
                 .description(request.getDescription())
                 .kakaoUrl(request.getKakaoUrl())
                 .projectType(request.getProjectType())
+                .projectMode(request.getProjectMode())
                 .totalMembers(request.getTotalMembers())
                 .duration(request.getDuration())
                 .deadline(request.getDeadline())
                 .startDate(request.getStartDate())
                 .techStacks(request.getTechStacks())
-                .projectMode(request.getProjectMode())
+                .teams(request.getTeams())
+                .requiredPositions(request.getRequiredPositions())
                 .build();
 
-        return projectRepository.save(project);
+        Project savedProject = projectRepository.save(project);
+
+        return ProjectDetailResponse.from(savedProject);
     }
 
     // 프로젝트 수정
-    // TODO: 프로젝트 모집글 수정은 글을 쓴 사람만이 할수 있게 해야 함
     @Transactional
-    public Project updateProject(Long projectId, ProjectUpdateRequest request) {
-        Project project = projectRepository.findById(projectId)
-                .orElseThrow(ProjectNotFoundException::new);
+    public void updateProject(Long projectId, ProjectUpdateRequest request) {
+        Project project = findProjectById(projectId);
+        validateMemberAccess(project);
 
         project.update(request);
-
-        return project;
     }
 
     // 프로젝트 삭제
-    // TODO: 프로젝트 모집글 삭제는 글을 쓴 사람만이 할수 있게 해야 함
     @Transactional
     public void deleteProject(Long projectId) {
-        Project project = projectRepository.findById(projectId)
-                .orElseThrow(ProjectNotFoundException::new);
+        Project project = findProjectById(projectId);
+        validateMemberAccess(project);
+
         projectRepository.delete(project);
     }
 
-    // 프로젝트 조회
-    public Project getProject(Long projectId) {
+    // 프로젝트 단일 조회
+    public ProjectDetailResponse getProjectById(Long projectId) {
+        Project project =  projectRepository.findById(projectId)
+                .orElseThrow(ProjectNotFoundException::new);
+
+        return ProjectDetailResponse.from(project);
+    }
+
+    // 모든 프로젝트 목록 조회
+    public List<ProjectSimpleResponse> getAllProjects() {
+        List<ProjectSimpleResponse> projects = projectRepository.findAll()
+                .stream()
+                .map(ProjectSimpleResponse::new)
+                .toList();
+
+        return projects;
+    }
+
+    ///////////// 편의성 메서드 ////////////
+    private void validateMemberAccess(Project project) {
+        Long currentUserId = getCurrentUserId();
+        if (!project.getProfile().getMember().getId().equals(currentUserId)) {
+            throw new UnauthorizedAccessException();
+        }
+    }
+
+    private Project findProjectById(Long projectId) {
         return projectRepository.findById(projectId)
                 .orElseThrow(ProjectNotFoundException::new);
     }
 
-    // 회원의 모든 프로젝트 조회 (회원이 등록한 모집글)
-    public List<Project> getProjectsByMember(Long memberId) {
-        return projectRepository.findProjectsByMemberId(memberId);
-    }
-
-    // TODO: 회원이 지원한 모든 지원글들을 조회
-    // ???
-
-    // 모든 프로젝트 목록 조회
-    public List<Project> getAllProjects() {
-        return projectRepository.findAll();
-    }
-
-    private Project getProjectByMemberId(Long memberId) {
-        return projectRepository.findByMemberId(memberId)
-                .orElseThrow(ProjectNotFoundException::new);
+    private Profile findProfileByMemberId(Long memberId) {
+        return profileRepository.findByMemberId(memberId)
+                .orElseThrow(ProfileNotFoundException::new);
     }
 
     private Long getCurrentUserId() {
